backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
alembic==1.12.1

# HTTP Client
httpx==0.25.1

# Celery
celery==5.3.4
redis==5.0.1

# LLM APIs
openai==1.3.3
google-generativeai==0.3.1

# Email
sendgrid==6.10.0
python-dateutil==2.8.2
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6 